refactor: split behavior table in separate module file
